==> 00_env_init.Rmd <==
---
title: "00_env_init"
author: "Puvvula"
date: "2024-04-15"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, stringr,  skimr,viridis, reshape2, kableExtra, stringr, tidygeocoder)
pacman::p_load(tidyverse, parallel, raster, doParallel, sf, lubridate)
library(tidyverse, appliedepi)
```

```{r}
data<- "~/Documents/peach_heat_metab/data/pj_data/"
analy_dat<- "~/Documents/peach_heat_metab/analy_data/"
result<- "~/Documents/peach_heat_metab/result/"
```
