==> 02_temperature_exp_assmt.Rmd <==
---
title: "02_temperature_exp_assmt"
author: "Puvvula"
date: "2024-04-18"
output: pdf_document
---

#read participant data
```{r}
df<- read_csv(paste0(analy_dat, "df_for_exposures.csv")) |>
  dplyr::select(-c(3:7, 9, 11:16, 18:20, 22, 33:37, 40)) |>
  mutate(lmp_date = as.Date(mdy(lmp_date), format="%Y-%m-%d"),
         delivery_date = as.Date(mdy(delivery_date), format="%Y-%m-%d"),
         start_date = lmp_date - days(14)) |>
  drop_na(9, 18, 19) |>
  rename(id=pt_id, lat= latitude, long = longitude) |>
  dplyr::select(c(7, 17:19, 9))
```


```{r}
for (i in seq(1, nrow(df), by = 5)) {
    # Get a subset of 5 observations
    subset_df <- df[i:min(i+4, nrow(df)), ]
    # Define the name for temp_exp variable
    temp_exp_name <- paste0("temp_exp_", ceiling(i/5))
    # Call the function with the subset
    assign(temp_exp_name, getRasterValueInRange(subset_df, input_folder_name = "/Volumes/Extreme Pro/tmax/daily"))
}
```

```{r}
# List all objects in the global environment
all_objects <- ls()

# Filter out dataframes that start with "temp_exp_"
temp_exp_dataframes <- all_objects[grep("^temp_exp_", all_objects)]

# Initialize an empty dataframe to store merged data
merged_temp_exp <- data.frame()

# Merge all temp_exp_ dataframes into a single dataframe
if (length(temp_exp_dataframes) > 0) {
  merged_temp_exp <- Reduce(function(x, y) merge(x, y, all = TRUE), lapply(temp_exp_dataframes, get))
}
```

```{r}
df_exprt<- left_join(df, merged_temp_exp, by="id")

save(df_exprt, file = paste0(analy_dat, "merged_temp_exp.rda"))
```

#further cleaning
```{r}
load(paste0(analy_dat, "merged_temp_exp.rda"))

#split liest to individual variable
df_new <- split_and_expand(df_exprt, "raster_value", "temp_w_")

#export clean data
write_csv(df_new, paste0(analy_dat, "pt_demo_exp_fin.csv"))
```

#exploring exposure data
```{r}
#processed final data
df<- read_csv(paste0(analy_dat, "pt_demo_exp_fin.csv"))

#additional covars received from Rita on May 08, 2024
df_addl<- read_csv(paste0(data, "CRIB_deog_update.csv")) |>
  dplyr::select(c(1:5))

#final merge
# 5 subjects with NA's replaced as less than college degree
df_update<- left_join(df, df_addl, by= "id") |>
  mutate(parity = ifelse(parity == 0, "nulliparus", "parus"),
         education = recode(education,
                            "Advanced Degree" = "Advanced Degree",
                            "College Graduate" = "College Graduate",
                            "Completed High school" = "Less than College Degree",
                            "GED_AA" = "Less than College Degree",
                            "Less than High School" = "Less than College Degree",
                            "NA" = "Less than College Degree"),
         education = if_else(is.na(education), "Less than College Degree", education)) |>
  rename(
    "LMP-2_weeks" = temp_w_1,
    "LMP-1_week" = temp_w_2
  ) 

#renaming labels
for (i in 26:66) {
    col_name <- names(df_update)[i]
    if (startsWith(col_name, "temp_w_")) {
        new_name <- paste0("Pregnancy_week_", as.numeric(str_extract(col_name, "\\d+")) - 2)
        names(df_update)[i] <- new_name
    }
}
```

#export data for metabolomics analysis
```{r}
save(df_update, file=paste0(analy_dat, "df_for_metabolomics.rda"))
```
